Update Chapter03.md
Made changes as discussed and agreed to on Monday July 8, 2019. Also made changes in rest of Chapter 3.1 based on the changes we had agreed upon for consistency.
I could not upload the latest figures (permissions issue) and would revisit the Figure Captions once Rabi uploads the two diagrams that Bernard provided.